backend/requirements.txt
alembic==1.12.0
blinker==1.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dicttoxml==1.7.16
docx==0.2.4
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fsspec==2024.2.0
greenlet==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lxml==5.1.0
Mako==1.2.4
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.29.0
python-docx==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.2
SQLAlchemy==2.0.22
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.1
transformers==4.38.2
typing_extensions==4.8.0
urllib3==2.2.1
Werkzeug==3.0.0
